Extract common code of Method objects
- remove unused universe from Invokable
